+++ silo-core/contracts/lib/SiloLendingLib.sol
@@ -162,7 +162,7 @@ library SiloLendingLib {
         address _protectedShareToken,
         address _collateralShareToken,
         address _borrower
-    ) public view returns (bool possible) {
+    ) internal view returns (bool possible) {
         // _borrower cannot have any collateral deposited
         possible = IShareToken(_protectedShareToken).balanceOf(_borrower) == 0
             && IShareToken(_collateralShareToken).balanceOf(_borrower) == 0;