+++ silo-core/contracts/lib/SiloSolvencyLib.sol
@@ -24,13 +24,25 @@ library SiloSolvencyLib {
     uint256 internal constant _PRECISION_DECIMALS = 1e18;
     uint256 internal constant _INFINITY = type(uint256).max;
 
+    function getOrderedConfigs(ISilo _silo, ISiloConfig _config, address _borrower)
+        external
+        view
+        returns (ISiloConfig.ConfigData memory collateralConfig, ISiloConfig.ConfigData memory debtConfig)
+    {
+        (collateralConfig, debtConfig) = _config.getConfigs(address(_silo));
+
+        if (!validConfigOrder(collateralConfig.debtShareToken, debtConfig.debtShareToken, _borrower)) {
+            (collateralConfig, debtConfig) = (debtConfig, collateralConfig);
+        }
+    }
+
     /// @dev check if config was given in correct order
     /// @return orderCorrect TRUE means that order is correct OR `_borrower` has no debt and we can not really tell
     function validConfigOrder(
         address _collateralConfigDebtShareToken,
         address _debtConfigDebtShareToken,
         address _borrower
-    ) external view returns (bool orderCorrect) {
+    ) internal view returns (bool orderCorrect) {
         uint256 debtShareTokenBalance = IShareToken(_debtConfigDebtShareToken).balanceOf(_borrower);
 
         return
@@ -49,7 +61,7 @@ library SiloSolvencyLib {
         address _borrower,
         ISilo.AccrueInterestInMemory _accrueInMemory,
         uint256 debtShareBalance
-    ) external view returns (bool) {
+    ) internal view returns (bool) {
         uint256 ltv = getLtv(
             _collateralConfig, _debtConfig, _borrower, ISilo.OracleType.Solvency, _accrueInMemory, debtShareBalance
         );
@@ -160,7 +172,7 @@ library SiloSolvencyLib {
         ISilo.OracleType _oracleType,
         ISilo.AccrueInterestInMemory _accrueInMemory,
         uint256 _debtShareBalance
-    ) public view returns (uint256 ltvInDp) {
+    ) internal view returns (uint256 ltvInDp) {
         if (_debtShareBalance == 0) return 0;
 
         LtvData memory ltvData = getAssetsDataForLtvCalculations(