+++ silo-core/test/foundry/echidna-findings/EchidnaSetup.sol
@@ -7,6 +7,7 @@ import {Strings} from "openzeppelin-contracts/utils/Strings.sol";
 import {ISiloConfig} from "silo-core/contracts/interfaces/ISiloConfig.sol";
 import {ISilo} from "silo-core/contracts/interfaces/ISilo.sol";
 import {IShareToken} from "silo-core/contracts/interfaces/IShareToken.sol";
+import {SiloLensLib} from "silo-core/contracts/lib/SiloLensLib.sol";
 
 import {MintableToken} from "../_common/MintableToken.sol";
 import {SiloLittleHelper} from "../_common/SiloLittleHelper.sol";
@@ -14,6 +15,8 @@ import {SiloFixture, SiloConfigOverride} from "../_common/fixtures/SiloFixture.s
 
 // setup must match what was set for `EchidnaE2E`
 contract EchidnaSetup is SiloLittleHelper, Test {
+    using SiloLensLib for ISilo;
+
     uint256 constant ACTORS_COUNT = 3;
     mapping (uint256 index => address actor) actors;