+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- add `SiloLens` to reduced Silo size
+
 ### Changed
 - change visibility of `total` mapping to public
 - ensure total getters returns values with interest